clean source 2025.08.23 18:40 JHE

+++ JAVA-ShoppingMall-team2-develop/21. MVCLayerDemo/src/com/shopping/controller/UserController.java
@@ -3,33 +3,32 @@ package com.shopping.controller;
 import java.util.List;
 import java.util.Scanner;
 
-import com.shopping.model.User;
 import com.shopping.Auth.Session;
 import com.shopping.model.Admin;
 import com.shopping.model.Order;
 import com.shopping.model.Role;
-import com.shopping.service.UserService;
-import com.shopping.service.AuthService;
-import com.shopping.service.OrderService;
+import com.shopping.model.User;
 import com.shopping.repository.DefaultFileOrderRepository;
 import com.shopping.repository.FileAdminRepository;
-import com.shopping.repository.FileOrderRepository;
 import com.shopping.repository.FileUserRepository;
+import com.shopping.service.AuthService;
+import com.shopping.service.OrderService;
+import com.shopping.service.UserService;
 
 
 /**
- * 사용자 관련 UI를 담당하는 컨트롤러 
+ * 사용자 관련 UI를 담당하는 컨트롤러
  * Presentation Layer의 일부로 사용자 입력을 받고 결과를 표시
  */
 public class UserController {
-	
+
 	private UserService userService;
 	private AuthService authService;
 	private Scanner scanner;
 	private OrderService orderService;
 	private Session session;
 
-	
+
 	public UserController(Session session) {
 	    FileUserRepository userRepo = new FileUserRepository();
 	    FileAdminRepository adminRepo = new FileAdminRepository();
@@ -40,19 +39,19 @@ public class UserController {
 	    this.scanner = new Scanner(System.in);
 	    this.session = session;
 	}
-	
+
 	public AuthService getAuthService() {
 	    return authService;
 	}
 
-	
+
 	// 마이페이지
 	public void myPage(User user) {
 		while(true) {
 			System.out.println("\n╔════════════════════════════════════════════╗");
 	        System.out.println("║                   마이 페이지                   ║");
 	        System.out.println("╚══════════════════════════════════════════════╝\n");
-	        
+
 	        System.out.println("║1. 내 정보 조회                                  ║");
 	        System.out.println("║2. 비밀번호 변경                                  ║");
 	        System.out.println("║3. 개인정보 수정                                  ║");
@@ -62,7 +61,7 @@ public class UserController {
 	        System.out.println("╚══════════════════════════════════════════════╝\n");
 
 	        String choice = scanner.nextLine();
-	        
+
 	        switch (choice) {
             case "1":
             	showUserInfo();
@@ -108,8 +107,8 @@ public class UserController {
 	        System.out.println("탈퇴 취소");
 	    }
 	}
-   
-	        
+
+
     // 주문 내역 조회
     private void viewOrderHistory() {
         User user = (User) authService.getLoggedInUser();
@@ -131,8 +130,8 @@ public class UserController {
                                ", 총 금액: " + order.getTotalPrice() +
                                ", 날짜: " + order.getOrderDate());
         }
-    }     
-	        
+    }
+
 	// 개인정보 수정
     private void editPersonalInfo() {
         User user = (User) authService.getLoggedInUser();
@@ -150,8 +149,8 @@ public class UserController {
         } catch (Exception e) {
             System.out.println("개인정보 수정 실패: " + e.getMessage());
         }
-    }        
-	        
+    }
+
 	// 비밀번호 변경
 	private void changePassword() {
 		User user = (User) authService.getLoggedInUser();
@@ -172,21 +171,21 @@ public class UserController {
 	// 회원가입 처리
 	public void register() {
 		System.out.println("\n== 회원가입 ==");
-		
+
 		// 아이디 입력 받기
 		System.out.print("아이디 (3자 이상, 영문/숫자): ");
 		String id = scanner.nextLine();
-		
+
 		// 입력 검증
 		if (id.length() < 3) {
 			System.out.println("아이디는 3자 이상이어야 합니다.");
 			return;
 		}
-		
+
 		// 패스워드 입력 받기
 		System.out.print("패스워드 (4자 이상): ");
 		String password = scanner.nextLine();
-		
+
 		if (password.length() < 4) {
 			System.out.println("패스워드는 4자 이상이어야 합니다.");
 			return;
@@ -195,36 +194,36 @@ public class UserController {
 		// 이메일 입력 받기
 		System.out.print("이메일 (아이디@도메인): ");
 		String email = scanner.nextLine();
-		
+
 		if (!email.contains("@") || !email.contains(".")) {
 			System.out.println("올바른 이메일 형식이 아닙니다.");
 			return;
 		}
-		
+
 		// 이름 입력 받기
 		System.out.print("이름: ");
 		String name = scanner.nextLine();
-		
+
 		if (name.trim().isEmpty()) {
 			System.out.println("이름을 입력해주세요.");
 			return;
 		}
-		
+
 		try {
 			User user = userService.register(id, password, email, name);
-			
+
 			System.out.println("회원가입 성공!");
 			System.out.println("환영합니다, " + user.getName() + "님!");
-			System.out.println("초기 잔액: " + (int)user.getBalance() + "원");			
+			System.out.println("초기 잔액: " + user.getBalance() + "원");
 		} catch (Exception e) {
 			System.out.println("회원가입 실패: " + e.getMessage());
 		}
 	}
-	
+
 	// 로그인 처리
 	public void login() {
 	    System.out.println("\n== 로그인 ==");
-	    
+
 	    if (session.isLoggedIn()) {
 	        System.out.println("이미 로그인되어 있습니다.");
 	        return;
@@ -234,11 +233,11 @@ public class UserController {
 	    String email = scanner.nextLine();
 	    System.out.print("패스워드: ");
 	    String password = scanner.nextLine();
-	    
+
 	    try {
 	        Role role = authService.login(email, password);  // Role 반환받기
 	        Object loggedUser = authService.getLoggedInUser();
-	        
+
 	        if (role == Role.USER && loggedUser instanceof User user) {
 	            session.login(user.getId(), Role.USER, user);
 	            System.out.println("로그인 성공! 회원: " + user.getName());
@@ -266,7 +265,7 @@ public class UserController {
 	}
 
 
-	
+
 	// 내 정보 보기
 	private void showUserInfo() {
 	    User user = session.getUser();  // 세션에서 User 객체 직접 읽기
@@ -278,7 +277,7 @@ public class UserController {
 	    System.out.println("ID: " + user.getId());
 	    System.out.println("이름: " + user.getName());
 	    System.out.println("이메일: " + user.getEmail());
-	    System.out.println("잔액: " + (int) user.getBalance() + "원");
+	    System.out.println("잔액: " + user.getBalance() + "원");
 	}
 
 }
